# aws/lambda/swagger.json
{"swagger": "2.0", "info": {"version": "1.0", "title": "secret"}, "schemes": ["https"], "paths": {"/": {"get": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}}}, "/{lang}": {"get": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}, "parameters": [{"name": "lang", "in": "path", "required": true, "type": "string"}]}}, "/decrypt/{id}": {"get": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}, "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}]}}, "/delete/{id}": {"delete": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}, "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}]}}, "/encrypt": {"post": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}}}, "/ping": {"get": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}}}, "/{lang}/view/{id}": {"get": {"consumes": ["application/json"], "produces": ["application/json"], "responses": {"200": {"description": "200 response", "schema": {"$ref": "#/definitions/Empty"}}}, "x-amazon-apigateway-integration": {"responses": {"default": {"statusCode": "200"}}, "uri": "${invoke_arn}", "passthroughBehavior": "when_no_match", "httpMethod": "POST", "contentHandling": "CONVERT_TO_TEXT", "type": "aws_proxy"}, "parameters": [{"name": "lang", "in": "path", "required": true, "type": "string"}, {"name": "id", "in": "path", "required": true, "type": "string"}]}}}, "definitions": {"Empty": {"type": "object", "title": "Empty Schema"}}, "x-amazon-apigateway-binary-media-types": ["application/octet-stream", "application/x-tar", "application/zip", "audio/basic", "audio/ogg", "audio/mp4", "audio/mpeg", "audio/wav", "audio/webm", "image/png", "image/jpg", "image/jpeg", "image/gif", "video/ogg", "video/mpeg", "video/webm"]}